add two-trigger sequencing
To disable a trigger, set its hitcount() to 0

Otherwise, trigger() condition has to clear before trigger2()
condition.

The coding is a bit janky -- had to wrangle with the stream
format and injecting "hit" overlays.

This code also /probably/ doesn't work with cd > 1. I've had
problems in general with getting litescope to work with cd > 1,
though.