core: Change upload protocol to allow bursting through xBone
Instead of using a very wide CSR, we force it to max 32 bits
and read each captured words as several sub-words.

Also, instead of checking for 'valid' flag every time, we
just read the 'level' of the memory buffer. Given the way
LiteScope works, we know capture is done and this is how many
words there is to read.

All in all this means that reading the data off the
buffer is just reading the same address over and over meaning
we can use very long bursts which helps _a_lot_ to speed things
up.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>